streaming/requirements.txt
apache-beam[gcp]
google-cloud-pubsub
numpy
pandas
tabulate
